app/requirements.txt
fastapi
uvicorn
langchain
unstructured
autogen
transformers
sqlalchemy
python-dotenv
langchain_core
langchain_community
llama-cpp-python
rich